renamed package to com.algorithmia.algorithm, and algoHandler to `Handler`
