+++ src/main/java/com/algorithmia/algorithm/RequestHandler.java
@@ -1,4 +1,4 @@
-package com.algorithmia.algorithmHandler;
+package com.algorithmia.algorithm;
 
 import com.google.gson.*;
 import org.apache.commons.codec.binary.Base64;
