+++ src/main/java/com/algorithmia/algorithm/Required.java
@@ -1,4 +1,4 @@
-package com.algorithmia.algorithmHandler;
+package com.algorithmia.algorithm;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
