+++ src/test/java/AlgorithmHandler/tests/AdvancedTests/InputTypeFailure.java
@@ -2,8 +2,8 @@ package AlgorithmHandler.tests.AdvancedTests;
 
 import AlgorithmHandler.algorithms.LoadingAlgorithm;
 import AlgorithmHandler.algorithms.MatrixAlgorithm;
-import AlgorithmHandler.tests.AlgorithmHandlerTestBase;
-import com.algorithmia.algorithmHandler.AlgorithmHandler;
+import AlgorithmHandler.tests.HandlerTestBase;
+import com.algorithmia.algorithm.Handler;
 import com.google.gson.Gson;
 import com.google.gson.JsonObject;
 import org.junit.Assert;
@@ -14,7 +14,7 @@ import java.io.InputStream;
 import java.nio.file.Files;
 import java.nio.file.Paths;
 
-public class InputTypeFailure extends AlgorithmHandlerTestBase {
+public class InputTypeFailure extends HandlerTestBase {
 
     private LoadingAlgorithm algo = new LoadingAlgorithm();
     private Gson gson = new Gson();
@@ -42,7 +42,7 @@ public class InputTypeFailure extends AlgorithmHandlerTestBase {
     @Test
     public void RunAlgorithm() throws Exception {
 
-        AlgorithmHandler handler = new AlgorithmHandler<>(algo.getClass(), algo::Apply, algo::DownloadModel);
+        Handler handler = new Handler<>(algo.getClass(), algo::Apply, algo::DownloadModel);
         InputStream fakeIn = new ByteArrayInputStream(request.toString().getBytes());
 
         System.setIn(fakeIn);
